# src/docs/swagger.yaml
basePath: /v1
definitions:
...
  httpapi.UserUpdateRequest:
    properties:
      email:
        type: string
      password:
        type: string
      role:
        type: string
    type: object
...
  users.CreateUserRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  users.UserResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      role:
        $ref: '#/definitions/users.UserRole'
    type: object
  users.UserRole:
    enum:
    - user
    - admin
    type: string
    x-enum-varnames:
    - RoleUser
    - RoleAdmin
info:
  contact: {}
  description: sniply_api HTTP API.
  title: sniply_api API
  version: "1.0"
paths:
...
  /users:
    get:
      parameters:
      - description: search
        in: query
        name: q
        type: string
      - description: limit
        in: query
        name: limit
        type: integer
      - description: offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/users.UserResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - SessionAuth: []
      - ApiKeyAuth: []
      summary: List users (admin)
      tags:
      - users
    post:
      consumes:
      - application/json
      parameters:
      - description: user
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/users.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/users.UserResponse'
        "400":
          description: Bad Request
          schema:
            type: string
        "409":
          description: Conflict
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Create user
      tags:
      - users
  /users/{id}:
    delete:
      parameters:
      - description: user id
        in: path
        name: id
        required: true
        type: string
      - description: CSRF token (required for SessionAuth)
        in: header
        name: X-CSRF-Token
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - SessionAuth: []
      - ApiKeyAuth: []
      summary: Delete user (admin or self)
      tags:
      - users
    put:
      consumes:
      - application/json
      parameters:
      - description: user id
        in: path
        name: id
        required: true
        type: string
      - description: user
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/httpapi.UserUpdateRequest'
      - description: CSRF token (required for SessionAuth)
        in: header
        name: X-CSRF-Token
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - SessionAuth: []
      - ApiKeyAuth: []
      summary: Update user (admin or self)
      tags:
      - users
  /users/me:
    delete:
      parameters:
      - description: CSRF token (required for SessionAuth)
        in: header
        name: X-CSRF-Token
        type: string
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - SessionAuth: []
      - ApiKeyAuth: []
      summary: Delete current user
      tags:
      - users
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/users.UserResponse'
        "401":
          description: Unauthorized
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - SessionAuth: []
      - ApiKeyAuth: []
      summary: Get current user
      tags:
      - users
    put:
      consumes:
      - application/json
      parameters:
      - description: user
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/httpapi.UserUpdateRequest'
      - description: CSRF token (required for SessionAuth)
        in: header
        name: X-CSRF-Token
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - SessionAuth: []
      - ApiKeyAuth: []
      summary: Update current user
      tags:
      - users
securityDefinitions:
  ApiKeyAuth:
    description: 'API key (X-API-Key or Authorization: Bearer)'
    in: header
    name: X-API-Key
    type: apiKey
  SessionAuth:
    description: Cookie header with sniply_session
    in: header
    name: Cookie
    type: apiKey
swagger: "2.0"